// src/shared/client/zod.gen.ts
export const zBookmark = z.object({
    id: z.string(),
    createdAt: z.string(),
    modifiedAt: z.union([
        z.string(),
        z.null()
    ]),
    title: z.union([
        z.string(),
        z.null()
    ]).optional(),
    archived: z.boolean(),
    favourited: z.boolean(),
    taggingStatus: z.enum([
        'success',
        'failure',
        'pending'
    ]),
    summarizationStatus: z.enum([
        'success',
        'failure',
        'pending'
    ]),
    note: z.union([
        z.string(),
        z.null()
    ]).optional(),
    summary: z.union([
        z.string(),
        z.null()
    ]).optional(),
    tags: z.array(z.object({
        id: z.string(),
        name: z.string(),
        attachedBy: z.enum([
            'ai',
            'human'
        ])
    })),
    content: z.union([
        z.object({
            type: z.enum([
                'link'
            ]),
            url: z.string(),
            title: z.union([
                z.string(),
                z.null()
            ]).optional(),
            description: z.union([
                z.string(),
                z.null()
            ]).optional(),
            imageUrl: z.union([
                z.string(),
                z.null()
            ]).optional(),
            imageAssetId: z.union([
                z.string(),
                z.null()
            ]).optional(),
            screenshotAssetId: z.union([
                z.string(),
                z.null()
            ]).optional(),
            fullPageArchiveAssetId: z.union([
                z.string(),
                z.null()
            ]).optional(),
            precrawledArchiveAssetId: z.union([
                z.string(),
                z.null()
            ]).optional(),
            videoAssetId: z.union([
                z.string(),
                z.null()
            ]).optional(),
            favicon: z.union([
                z.string(),
                z.null()
            ]).optional(),
            htmlContent: z.union([
                z.string(),
                z.null()
            ]).optional(),
            crawledAt: z.union([
                z.string(),
                z.null()
            ]).optional(),
            author: z.union([
                z.string(),
                z.null()
            ]).optional(),
            publisher: z.union([
                z.string(),
                z.null()
            ]).optional(),
            datePublished: z.union([
                z.string(),
                z.null()
            ]).optional(),
            dateModified: z.union([
                z.string(),
                z.null()
            ]).optional()
        }),
        z.object({
            type: z.enum([
                'text'
            ]),
            text: z.string(),
            sourceUrl: z.union([
                z.string(),
                z.null()
            ]).optional()
        }),
        z.object({
            type: z.enum([
                'asset'
            ]),
            assetType: z.enum([
                'image',
                'pdf'
            ]),
            assetId: z.string(),
            fileName: z.union([
                z.string(),
                z.null()
            ]).optional(),
            sourceUrl: z.union([
                z.string(),
                z.null()
            ]).optional(),
            size: z.union([
                z.number(),
                z.null()
            ]).optional(),
            content: z.union([
                z.string(),
                z.null()
            ]).optional()
        }),
        z.object({
            type: z.enum([
                'unknown'
            ])
        })
    ]),
    assets: z.array(z.object({
        id: z.string(),
        assetType: z.enum([
            'screenshot',
            'assetScreenshot',
            'bannerImage',
            'fullPageArchive',
            'video',
            'bookmarkAsset',
            'precrawledArchive',
            'linkHtmlContent',
            'unknown'
        ])
    }))
});